requirements.txt
Flask==3.1.2
Werkzeug==3.1.4
Jinja2==3.1.6
click==8.3.1
itsdangerous==2.2.0
MarkupSafe==3.0.3
Pillow==10.4.0
reportlab==4.4.6
gunicorn==23.0.0
